fix: download de PDF na páginas de documentos do estágio
- Move GET handler de download/route.ts para [docId]/route.ts (Next 16 + Turbopack ignorava rota com 2 segmentos dinâmicos + literal)
- Remove ficheiro download/route.ts
- Remove /download do path nos URLs do frontend
- fetchPdfBytes no pdf-viewer só tenta proxy para URLs absolutas do Firebase Storage
