+++ .gitignore
@@ -58,3 +58,4 @@ data/
 /functions/lib
 package-lock.json
 pnpm-lock.yaml
+*.stackdump
